feat: add details (with styles)
